show agree console when delete directories

+++ lib/houki/executors/directory.rb
@@ -1,5 +1,6 @@
 require 'fileutils'
 require 'terminal-table'
+require 'highline/import'
 
 module Houki::Executors
   class Directory
@@ -8,9 +9,12 @@ module Houki::Executors
     def self.run(directories)
       table = Terminal::Table.new(title: 'Directories', rows: directories.combination(1).to_a)
       $stdout.puts table
-      directories.each do |directory|
-        root_dirs = Dir.entries(directory).reject { |f| EXCLUDED_DIRS.include?(f) }
-        FileUtils.rm_r(root_dirs.collect { |f| File.join(directory, f) })
+
+      if agree('Are you sure you want to delete directories? ')
+        directories.each do |directory|
+          root_dirs = Dir.entries(directory).reject { |f| EXCLUDED_DIRS.include?(f) }
+          FileUtils.rm_r(root_dirs.collect { |f| File.join(directory, f) })
+        end
       end
     rescue Errno::ENOENT
     end
